Remove dependency on R2API + Rename folder
