+++ RoR2Mods.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31911.196
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActuallyFaster", "ExamplePlugin\ActuallyFaster.csproj", "{2F8A40CC-6D15-486D-B689-0A7A472DB89E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ActuallyFaster", "ActuallyFaster\ActuallyFaster.csproj", "{2F8A40CC-6D15-486D-B689-0A7A472DB89E}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Metadata", "Metadata", "{3A26DDA8-2697-4D64-962B-DB9E2DC00EF6}"
 	ProjectSection(SolutionItems) = preProject
