+++ manifest.json
@@ -1,6 +1,6 @@
 {
     "name": "Actually_Faster",
-    "version_number": "1.0.8",
+    "version_number": "1.0.9",
     "website_url": "https://github.com/FelixVolo/ActuallyFaster",
     "description": "A mod that makes the printer, scrapper, chance shrine and cauldrons go brrrrrt. This one actually works!",
     "dependencies": [
